Refactor: 레티나 디스플레이 대응하기 위해 이미지 png -> svg로 변경
