+++ src/components/CommentBar/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import user_img_small from "../../assets/images/user_img_small.png";
+import user_img_small from "../../assets/images/user_img_small.svg";
 
 const CommentContainer = styled.div`
   position: fixed;
